Importing meshes is available now, currently set to demo in windows
Can import 3d models using assimp into a collection of separate meshes (a Scene) or into a single mesh (a Mesh).

create, load, update, then setupbuffer to use it.

Currently have to manually move the shaders folder, models folder and the assimp-vc143-mtd.dll (from running the assimp cmake) into the folder our menv.exe is in. Perhaps write a shell and bash script to automate this or find a way to do it with the CMakeList file.